docker support

+++ scripts/cleanup.sh
@@ -9,9 +9,13 @@ apt-get update
 # Delete unneeded files.
 rm -f /home/vagrant/*.sh
 
-# Zero out the rest of the free space using dd, then delete the written file.
-dd if=/dev/zero of=/EMPTY bs=1M
-rm -f /EMPTY
+# don't zero disk for docker build
+if [ ! -f /.dockerenv ]; then
+	# Zero out the rest of the free space using dd, then delete the written file.
+	dd if=/dev/zero of=/EMPTY bs=1M
+	rm -f /EMPTY
 
-# Add `sync` so Packer doesn't quit too early, before the large file is deleted.
-sync
+	# Add `sync` so Packer doesn't quit too early, before the large file is deleted.
+	sync
+fi
+
